Relocate hook
