+++ website/src/app/[lang]/[region]/(website)/me/hooks.ts
@@ -6,7 +6,6 @@ import { DONATION_CERTIFICATE_FIRESTORE_PATH } from '@socialincome/shared/src/ty
 import { Employer, EMPLOYERS_FIRESTORE_PATH } from '@socialincome/shared/src/types/employers';
 import { USER_FIRESTORE_PATH } from '@socialincome/shared/src/types/user';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { Geo } from '@vercel/functions';
 import { addDoc, collection, deleteDoc, doc, getDocs, query, Timestamp, updateDoc, where } from 'firebase/firestore';
 import { useContext } from 'react';
 import { useFirestore } from 'reactfire';
@@ -170,22 +169,4 @@ export const useAddEmployer = () => {
 		});
 		await queryClient.invalidateQueries({ queryKey: ['me', 'employers'] });
 	};
-};
-
-export const useGeolocation = () => {
-	const api = useApi();
-	const {
-		data: geolocation,
-		isLoading,
-		error,
-	} = useQuery({
-		queryKey: ['geolocation'],
-		queryFn: async () => {
-			const response = await api.get('/api/geolocation');
-			return (await response.json()) as Geo;
-		},
-		staleTime: Infinity,
-	});
-
-	return { geolocation, isLoading, error };
 };
